requirements.txt
# 核心功能依赖
opencv-python>=4.5.0
numpy>=1.20.0
kafka-python>=2.0.0
ultralytics>=8.0.0
python-dotenv>=0.19.0

# 用于Kafka
avro>=1.10.0

# 用于图像处理和显示
Pillow>=8.0.0
matplotlib>=3.5.0

# 用于环境管理
python-dotenv>=0.19.0